requirements.txt
flask
flask_smorest
flask-sqlalchemy
flask-dotenv
python-dotenv
marshmallow
Flask-Babel

